fix: improve some moments according to review
